fix(layout): ensure screen reconfigure check happens in layout()
Previously, the logic was in `Layout.show()`, but `Layout.show()` isn't
invoked if screens are reconfigured but the group remained visible on
one of the screens. Only until it was swapped out for another group and
switched back, does `show()` get invoked, where our resize logic would
finally be invoked.